consolidate gotoTime into module
