+++ lib/doppler/action.js
@@ -6,6 +6,7 @@ const tAtX = require('./t-at-x.js');
 const refreshCursor = require('./refresh-cursor.js');
 const _ = require('lodash');
 const getX = require('./get-x.js');
+const gotoTime = require('./goto-time.js');
 
 const createBookmarkAtCursor = (event, content, cursor, pstate, ws, render) => {
   // Create a new bookmark from the PV.
@@ -183,20 +184,11 @@ const deleteMarker = (event, content, cursor, pstate, ws, render) => {
   return true;
 };
 
-const test = (event, content, cursor, pstate, ws, render) => {
-  // A place to put code to test so you can test by hitting the t key.
-  return true;
-};
-
-function isDigit(chr) {
-      return  chr.match(/[0-9]/i);
-}
-
 let timeStr = "";
 const userGotoTime = (event, content, cursor, pstate, ws, render) => {
   // A place to put code to test so you can test by hitting the t key.
   console.log(event);
-  if (isDigit(event.key)) {
+  if (event.key.match(/[0-9]/i)) {
     timeStr += event.key;
     console.log(timeStr);
   }
@@ -213,32 +205,11 @@ const userGotoTime = (event, content, cursor, pstate, ws, render) => {
   return true;
 };
 
-var isNumber = function isNumber(value) {
-  const n = parseInt(value);
-  return typeof n === 'number' && isFinite(n);
+const test = (event, content, cursor, pstate, ws, render) => {
+  // A place to put code to test so you can test by hitting the t key.
+  return true;
 };
 
-var gotoTime = function gotoTime(time, pstate, cursor, centerCursor, render) {
-  //console.log('gotoTime ' + time);
-  const xStartExact = tAtX(pstate.sidebarWidth, pstate);
-  const xFinishExact = tAtX(pstate.width, pstate);
-  const midOfs = (xFinishExact - xStartExact) / 2;
-
-  if (time == -1 || !isNumber(time)) {
-    return true;
-  }
-
-  const isVisible = xStartExact <= time && time <= xFinishExact;
-  //console.log(isVisible, centerCursor);
-  if (!isVisible || centerCursor == 'true') {
-    // If not visible, reposition the display to put the target in the center
-    pstate.xOffset = xOffsetAtT(time - midOfs, pstate);
-  }
-
-  render();
-  pstate.xCursor = getX(pstate, time);
-  refreshCursor(cursor, pstate, true);
-};
 
 module.exports = {
   i: createBookmarkAtCursor,
