+++ lib/ws-endpoint.js
@@ -7,7 +7,7 @@ const timeAtCursor = require('./doppler/time-at-cursor.js');
 const refreshCursor = require('./doppler/refresh-cursor.js');
 const scale = require('./doppler/scale.js');
 const _ = require('lodash');
-const instsAtTime = require('./doppler/instructions-at-time.js');
+const gotoTime = require('./doppler/goto-time.js');
 
 function finish(render, cursor, pstate) {
   render();
@@ -26,33 +26,6 @@ const ws = (function () {
 
   websocket.onopen = () => {
     console.log('websocket: ' + websocket.url);
-  };
-
-  var isNumber = function isNumber(value) {
-    const n = parseInt(value);
-    return typeof n === 'number' && isFinite(n);
-  };
-
-  var gotoTime = function gotoTime(time, pstate, cursor, centerCursor) {
-    //console.log('gotoTime ' + time);
-    const xStartExact = tAtX(pstate.sidebarWidth, pstate);
-    const xFinishExact = tAtX(pstate.width, pstate);
-    const midOfs = (xFinishExact - xStartExact) / 2;
-
-    if (time == -1 || !isNumber(time)) {
-      return true;
-    }
-
-    const isVisible = xStartExact <= time && time <= xFinishExact;
-    //console.log(isVisible, centerCursor);
-    if (!isVisible || centerCursor == 'true') {
-      // If not visible, reposition the display to put the target in the center
-      pstate.xOffset = xOffsetAtT(time - midOfs, pstate);
-    }
-
-    render();
-    pstate.xCursor = getX(pstate, time);
-    refreshCursor(cursor, pstate, true);
   };
 
   websocket.onmessage = (msg) => {
@@ -94,7 +67,7 @@ const ws = (function () {
 
       case 'centercursor': {
         const time = tAtX(pstate.xCursor, pstate);
-        gotoTime(time, pstate, cursor, true);
+        gotoTime(time, pstate, cursor, true, render);
         return;
       }
 
@@ -129,7 +102,7 @@ const ws = (function () {
         }
         return;
       case 'goto': {
-        gotoTime(cmdArray[1], pstate, cursor, cmdArray[2]);
+        gotoTime(cmdArray[1], pstate, cursor, cmdArray[2], render);
         return true;
       }
       case 'close':
